use INT_MIN/MAX for 32bit ranges (64bit ticket #6)
git-svn-id: http://svn.osgeo.org/fdo/trunk@2512 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Fdo/Unmanaged/Src/Fdo/Parse/Lex.cpp
@@ -877,7 +877,7 @@ bool FdoLex::getnumber(FdoParse *pParse, bool sign)
         if ( checkStr == str ) {
             // String converted to in64 ok.
 
-            if (i64num <= LONG_MAX && i64num >= LONG_MIN) 
+            if (i64num <= INT_MAX && i64num >= INT_MIN) 
 		    {
                 // In Int32 range so wrap in int32
 			    try
